refactor(dashboard): isolate backend recovery boundary
Issue: PAN-3373
