+++ src/dashboard/frontend/src/App/AppRoutes.tsx
@@ -31,6 +31,7 @@ import { WorkspaceView } from '../components/workspace/WorkspaceView';
 import type { Tab } from '../components/Header';
 import type { Issue } from '../types';
 import type { ViewMode as ConversationViewMode } from '../components/chat/ConversationPanel';
+import { BackendConnectionBoundary } from './BackendConnectionBoundary';
 
 export interface PendingConversationTarget {
   conversationName: string;
@@ -44,6 +45,8 @@ type SelectProjectHandler = (projectName: string | null, opts?: { updateUrl?: bo
 
 interface AppRoutesProps {
   activeTab: Tab;
+  backendDown: boolean;
+  restarting: boolean;
   issues: Issue[];
   selectedConvId: string | null;
   conversationViewMode: ConversationViewMode;
@@ -75,6 +78,8 @@ interface AppRoutesProps {
 
 export function AppRoutes({
   activeTab,
+  backendDown,
+  restarting,
   issues,
   selectedConvId,
   conversationViewMode,
@@ -102,7 +107,7 @@ export function AppRoutes({
   keyboardShortcutsDisabled = false,
 }: AppRoutesProps) {
   return (
-    <>
+    <BackendConnectionBoundary backendDown={backendDown} restarting={restarting}>
       {activeTab === 'home' && (
         <div className="w-full h-full overflow-hidden">
           <HomeSwitch
@@ -271,6 +276,6 @@ export function AppRoutes({
           <DeaconActivityView />
         </div>
       )}
-    </>
+    </BackendConnectionBoundary>
   );
 }
